Commands manager refactor
